纹理透视映射，Z_buffer
修复纹理映射的偏差问题（忘记将w透视校正导致），添加Z_buffer缓冲区

+++ Direct3DGame/Direct3DGame/Texture.cpp
@@ -29,7 +29,7 @@ Texture::Texture(string path)
 
 AColor Texture::get_color(float u,float v)
 {
-	return pixels[(width-1)*u][(height-1)*v];
+	return pixels[(int)((width-1)*u)][(int)((height-1)*v)];
 }
 
 AColor Texture::get_color2(int u,int v)
